chore: Move core tests from reactive to the new core.test project
