+++ src/Directory.Packages.props
@@ -7,10 +7,10 @@
 		<PackageVersion Include="MSTest.TestAdapter" Version="2.2.9" />
 		<PackageVersion Include="MSTest.TestFramework" Version="2.2.9" />
 		<PackageVersion Include="Microsoft.Bcl.AsyncInterfaces" Version="6.0.0" />
-		<PackageVersion Include="Microsoft.CodeAnalysis.CSharp" Version="4.0.1" />
-		<PackageVersion Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.0.1" />
-		<PackageVersion Include="Microsoft.CodeAnalysis.Common" Version="4.0.1" />
-		<PackageVersion Include="Microsoft.CodeAnalysis.Workspaces.Common" Version="4.0.1" />
+		<PackageVersion Include="Microsoft.CodeAnalysis.CSharp" Version="4.2.0" />
+		<PackageVersion Include="Microsoft.CodeAnalysis.CSharp.Workspaces" Version="4.2.0" />
+		<PackageVersion Include="Microsoft.CodeAnalysis.Common" Version="4.2.0" />
+		<PackageVersion Include="Microsoft.CodeAnalysis.Workspaces.Common" Version="4.2.0" />
 		<PackageVersion Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
 		<PackageVersion Include="Microsoft.Extensions.DependencyInjection" Version="6.0.0" />
 		<PackageVersion Include="Microsoft.Extensions.Hosting" Version="6.0.1" />
