+++ src/Uno.Extensions.Core.Tests/Utils/GenerationTestHelper.cs
