+++ src/Uno.Extensions.Core/Edition/_AttributesVisibility.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Uno.Extensions.Edition;
+
+// As **Attribute.cs files are linked in the Core.Generators project (to ease code gen)
+// and the Core.Tests project is referencing both Core and Core.Generators projects,
+// attributes are present twice.
+
+// To resolve this, we don't set any access modifier in **Attributes.cs files (so they are internal by default),
+// and then make them public only in the Core project.
+
+public partial class DefaultConstructorAttribute
+{
+}
